goldfish_dma: directly accessing host memory
Video playback can really stress out the pipe.

This CL extends the pipe to allow mmap() into a
physically contiguous, large region. Writes
to this region appear "immediately" on the host,
requiring no further signals.

This is currently mainly used for speeding up
video playback on the emulator, but a general,
simple interface is presented:

- ioctls specify size of the region and allow
  lock/unlock() prevent concurrent access
- mmap() allocates the region for a guest pipe user

Change-Id: I98fcf0de468003d2639516dc296eb6221b7f0d0c
Signed-off-by: Lingfeng Yang <lfy@google.com>