mac80211_hwsim: set timestamps relative to boot
The Linux API specifies that the timestamp in probes and beacons should
be in microseconds since epoch. Android expects microseconds since boot.
This changes the behavior in mac80211_hwsim to match the expected
Android behavior. This is a short term solution to work around this
problem. Newer kernels have been fixed to provide information about time
since boot (in 6e19bc4b7091ffd26586100eee78232b44427ec7) and will not
need this workaround.

Change-Id: I487e71066293c255ed6763d7803da5943d4ae3ac
Signed-off-by: Bjoern Johansson <bjoernj@google.com>