Hook implementation for open syscall trracing
Implement a hook function that traces the open (or another) syscall. The function can be simplty attached to any syscall in order to log its behaviour.

The signature of the hook is
hook(syscall_name, argument_descriptor, argument_1, ..)
where
argument_descriptor = a string describing the format and order of the arguments. Currently the following arguments can be passed:
'd' - file descriptor, expects an integer
'n' - file name, expects a string
'p' - file path, expects a string
'f' - flags, expects an integer

The output is displayed in the console, and can be redirected to any file by dmesg > echo > filename.txt.

Sys_open_logger is a draft - to be used to synch the print of the logs in a file